+++ app/src/test/java/io/jim/tesserapp/util/util.kt
@@ -4,7 +4,7 @@ import io.jim.tesserapp.math.Vector
 import junit.framework.Assert.assertEquals
 
 @Suppress("unused")
-fun assertEquals(expected: Vector, actual: Vector, delta: Double) {
+fun assertEquals(expected: Vector, actual: Vector, delta: Float) {
     assertEquals(expected.x, actual.x, delta)
     assertEquals(expected.y, actual.y, delta)
     assertEquals(expected.z, actual.z, delta)
@@ -12,7 +12,7 @@ fun assertEquals(expected: Vector, actual: Vector, delta: Double) {
 }
 
 @Suppress("unused")
-fun assertEquals(message: String, expected: Vector, actual: Vector, delta: Double) {
+fun assertEquals(message: String, expected: Vector, actual: Vector, delta: Float) {
     assertEquals(message, expected.x, actual.x, delta)
     assertEquals(message, expected.y, actual.y, delta)
     assertEquals(message, expected.z, actual.z, delta)
